Added interfaces and ignored json data file

+++ apps/balance-service/src/balances/balances.controller.ts
@@ -3,6 +3,7 @@ import { BalancesService } from './balances.service';
 import { CreateBalanceDto } from './dto/create-balance.dto';
 import logger from '@app/shared/logger/winston-logger';
 import { InternalServerException } from 'libs/error-handling/exceptions/internal-server.exception';
+import { BadRequestException } from 'libs/error-handling/exceptions/bad-request.exception';
 
 @Controller('balances')
 export class BalancesController {
@@ -14,11 +15,14 @@ export class BalancesController {
     @Body() body: { asset: string; amount: number },
   ) {
     try {
-      const newBalance = await this.balancesService.addBalance(new CreateBalanceDto(userId, body.asset, body.amount));
-      return newBalance;
+      if (!body.asset || body.asset.length === 0) {
+        throw new BadRequestException('Missing asset or asset is empty');
+      }
+      await this.balancesService.addBalance(new CreateBalanceDto(userId, body.asset, body.amount));
+      return { message: `Balance ${body.asset} added successfully` };
     } catch (error) {
       logger.error('Error adding balance:', error);
-      throw new InternalServerException('Error adding balance');
+      throw new InternalServerException(error);
     }
   }
 
@@ -29,11 +33,14 @@ export class BalancesController {
     @Query('asset') asset: string,
   ) {
     try {
+      if (!asset || asset.length === 0) {
+        throw new BadRequestException('Missing asset or asset is empty');
+      }
       await this.balancesService.removeBalance(userId, asset, +body.amount);
       return { message: 'Balance removed successfully' };
     } catch (error) {
       logger.error('Error removing balance:', error);
-      throw new InternalServerException('Error removing balance');
+      throw new InternalServerException(error);
     }
   }
 
@@ -43,17 +50,20 @@ export class BalancesController {
       return await this.balancesService.getAllBalances();
     } catch (error) {
       logger.error('Error retrieving all balances:', error);
-      throw new InternalServerException('Error retrieving all balances');
+      throw new InternalServerException(error);
     }
   }
 
   @Get('userBalance')
   async getUserBalance(@Headers('X-User-ID') userId: string) {
     try {
+      if (!userId || userId.length === 0) {
+        throw new BadRequestException('Missing userId or userId is empty');
+      }
       return await this.balancesService.getAllUserBalances(userId);
     } catch (error) {
       logger.error(`Error retrieving balances for user ${userId}:`, error);
-      throw new InternalServerException('Error retrieving balances for user');
+      throw new InternalServerException(error);
     }
   }
 
@@ -63,10 +73,13 @@ export class BalancesController {
     @Query('currency') targetCurrency: string,
   ) {
     try {
+      if (!targetCurrency || targetCurrency.length === 0) {
+        throw new BadRequestException('Missing currency or currency is empty');
+      }
       return await this.balancesService.getUserBalancesInCurrency(userId, targetCurrency);
     } catch (error) {
       logger.error(`Error retrieving total balance for user ${userId} in ${targetCurrency}:`, error);
-      throw new InternalServerException('Error retrieving total balance');
+      throw new InternalServerException(error);
     }
   }
 }
